requirements.txt
pandas: 1.5.2
selenium: 4.16.0
beautifulsoup4: 4.12.2
tqdm: 4.65.0
pyyaml: 6.0.1